+++ src/instaparse/combinators.clj
@@ -33,3 +33,4 @@
 (defclone hide-tag c/hide-tag)
 
 (defclone ebnf cfg/ebnf)
+