+++ src/instaparse/combinators_source.clj
@@ -6,19 +6,23 @@
 
 (def Epsilon {:tag :epsilon})
 
-(defn opt [parser] 
+(defn opt "Optional, i.e., parser?"
+  [parser] 
   (if (= parser Epsilon) Epsilon
     {:tag :opt :parser parser}))
 
-(defn plus [parser]
+(defn plus "One or more, i.e., parser+"
+  [parser]
   (if (= parser Epsilon) Epsilon
     {:tag :plus :parser parser}))
 
-(defn star [parser] 
+(defn star "Zero or more, i.e., parser*"
+  [parser] 
   (if (= parser Epsilon) Epsilon
     {:tag :star :parser parser}))
 
-(defn alt [& parsers] 
+(defn alt "Alternation, i.e., parser1 | parser2 | parser3 | ..."
+  [& parsers] 
   (cond
     (every? (partial = Epsilon) parsers) Epsilon
     (singleton? parsers) (first parsers)
@@ -33,33 +37,47 @@
     ;(alt parser1 (cat (neg parser1) parser2))))
     {:tag :ord :parser1 parser1 :parser2 parser2}))
 
-(defn ord [& parsers]
+(defn ord "Ordered choice, i.e., parser1 / parser2"
+  [& parsers]
   (if (seq parsers)
     (ord2 (first parsers) (apply ord (rest parsers)))
     Epsilon))
 
-(defn cat [& parsers]
+(defn cat "Concatenation, i.e., parser1 parser2 ..."
+  [& parsers]
   (if (every? (partial = Epsilon) parsers) Epsilon
     (let [parsers (remove #{Epsilon} parsers)]
       (if (singleton? parsers) (first parsers) ; apply vector reduction
         {:tag :cat :parsers parsers}))))
 
-(defn string [s] 
+(defn string "Create a string terminal out of s" 
+  [s] 
   (if (= s "") Epsilon
     {:tag :string :string s}))
 
-(defn regexp [r]
+(defn regexp "Create a regexp terminal out of regular expression r"
+  [r]
   (let [s (str \^ r)]
     (if (= s "^") Epsilon
       {:tag :regexp :regexp (re-pattern s)})))
 
-(defn nt [s] {:tag :nt :keyword s})
+(defn nt "Refers to a non-terminal defined by the grammar map"
+  [s] 
+  {:tag :nt :keyword s})
 
-(defn look [parser] {:tag :look :parser parser}) 
+(defn look "Lookahead, i.e., &parser" 
+  [parser] 
+  {:tag :look :parser parser}) 
 
-(defn neg [parser] {:tag :neg :parser parser})
+(defn neg "Negative lookahead, i.e., !parser"
+  [parser] 
+  {:tag :neg :parser parser})
 
-(defn hide [parser] (assoc parser :hide true))
+(defn hide "Hide the result of parser, i.e., <parser>"
+  [parser] 
+  (assoc parser :hide true))
 
-(defn hide-tag [parser]
+(defn hide-tag "Hide the tag associated with this right-hand-side.  
+  Wrap this combinator around the entire right-hand side."  
+  [parser]
   (red parser raw-non-terminal-reduction))